Move test-utils to a separate module
