+++ settings.gradle
@@ -17,9 +17,11 @@ rootProject.name = 'mime'
 include ':checkstyle'
 include ':mime-native'
 include ':mime-ballerina'
+include ':mime-test-utils'
 
 project(':checkstyle').projectDir = file("build-config${File.separator}checkstyle")
 project(':mime-native').projectDir = file('native')
+project(':mime-test-utils').projectDir = file('test-utils')
 project(':mime-ballerina').projectDir = file('ballerina')
 
 gradleEnterprise {
